Refactor transport and player handshake
